refactor(app): extract HubRoutes from App.tsx (#316 phase 7)
Hub world, world map, casino, theatre, the two MiniGamesMenu entry points and
hub fishing.

`hubworld` and `location` stay a single HubWorld render gated on hubData —
the town comes from currentLocationKey, not the screen, and hubData is the
~1.5MB lazy fetch that only happens once the player heads for the hub. That
guard is carried over verbatim.

AppContextValue gains the hub state (hubData, currentLocationKey, worldMapKey,
restrictedTownNodeIds, the two minigame entry points), the setters HubWorld
navigates through, and goToWorldLocation / handleWorldBattle /
handleCampaign / handleCampaign2 / handleEndless.

App.tsx: 3,224 -> 3,118 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
